[INPUT] Rewrite from TS to JS 🤟
