+++ semcore/input/__tests__/__image_snapshots__/index-test-js-input-input-addon-should-support-interactive-prop-1-snap.png
