+++ semcore/input/__tests__/__image_snapshots__/index-test-js-input-should-support-input-addon-1-snap.png
