+++ semcore/input/__tests__/__image_snapshots__/index-test-js-input-should-support-placeholder-color-1-snap.png
