+++ semcore/input/__tests__/__image_snapshots__/index-test-js-input-should-support-states-1-snap.png
