+++ semcore/input/src/index.js
@@ -1,2 +1,1 @@
 export { default } from './Input';
-export * from './Input';
